Refactor into separate previewers
