+++ src/nl/hannahsten/texifyidea/action/preview/ShowEquationPreview.kt
@@ -6,7 +6,6 @@ import com.intellij.openapi.util.Key
 import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.psi.PsiElement
 import nl.hannahsten.texifyidea.TexifyIcons
-import nl.hannahsten.texifyidea.ui.PreviewFormUpdater
 import nl.hannahsten.texifyidea.util.findOuterMathEnvironment
 
 /**
